db: migrate access table to use GORM (#6371)
