+++ internal/cmd/serv.go
@@ -226,11 +226,12 @@ func runServ(c *cli.Context) error {
 				fail("Internal error", "Failed to get user by key ID '%d': %v", key.ID, err)
 			}
 
-			mode, err := db.UserAccessMode(user.ID, repo)
-			if err != nil {
-				fail("Internal error", "Failed to check access: %v", err)
-			}
-
+			mode := db.Perms.AccessMode(user.ID, repo.ID,
+				db.AccessModeOptions{
+					OwnerID: repo.OwnerID,
+					Private: repo.IsPrivate,
+				},
+			)
 			if mode < requestMode {
 				clientMessage := _ACCESS_DENIED_MESSAGE
 				if mode >= db.AccessModeRead {
@@ -333,10 +334,18 @@ func secureGitAnnex(path string, user *db.User, repo *db.Repository) error {
 	}
 	mode := db.AccessModeNone
 	if user != nil {
-		mode, err = db.UserAccessMode(user.ID, repo)
-		if err != nil {
-			fail("Internal error", "Fail to check access: %v", err)
-		}
+		// Code before ORM migration; check commit 2eaf1d693
+		// mode, err = db.UserAccessMode(user.ID, repo)
+		// if err != nil {
+		//	fail("Internal error", "Fail to check access: %v", err)
+		//}
+		// Code after ORM migration
+		mode = db.Perms.AccessMode(user.ID, repo.ID,
+			db.AccessModeOptions{
+				OwnerID: repo.OwnerID,
+				Private: repo.IsPrivate,
+			},
+		)
 	}
 	if mode < db.AccessModeWrite {
 		err = os.Setenv("GIT_ANNEX_SHELL_READONLY", "True")
