+++ internal/db/db.go
@@ -141,7 +141,7 @@ func openDB(opts conf.DatabaseOpts, cfg *gorm.Config) (*gorm.DB, error) {
 //
 // NOTE: Lines are sorted in alphabetical order, each letter in its own line.
 var Tables = []interface{}{
-	new(AccessToken),
+	new(Access), new(AccessToken),
 	new(LFSObject), new(LoginSource),
 }
