+++ internal/db/perms.go
@@ -14,9 +14,9 @@ import (
 // NOTE: All methods are sorted in alphabetical order.
 type PermsStore interface {
 	// AccessMode returns the access mode of given user has to the repository.
-	AccessMode(userID int64, repo *Repository) AccessMode
+	AccessMode(userID, repoID int64, opts AccessModeOptions) AccessMode
 	// Authorize returns true if the user has as good as desired access mode to the repository.
-	Authorize(userID int64, repo *Repository, desired AccessMode) bool
+	Authorize(userID, repoID int64, desired AccessMode, opts AccessModeOptions) bool
 	// SetRepoPerms does a full update to which users have which level of access to given repository.
 	// Keys of the "accessMap" are user IDs.
 	SetRepoPerms(repoID int64, accessMap map[int64]AccessMode) error
@@ -24,19 +24,72 @@ type PermsStore interface {
 
 var Perms PermsStore
 
+// Access represents the highest access level of a user has to a repository.
+// The only access type that is not in this table is the real owner of a repository.
+// In case of an organization repository, the members of the owners team are in this table.
+type Access struct {
+	ID     int64
+	UserID int64      `xorm:"UNIQUE(s)" gorm:"uniqueIndex:access_user_repo_unique;NOT NULL"`
+	RepoID int64      `xorm:"UNIQUE(s)" gorm:"uniqueIndex:access_user_repo_unique;NOT NULL"`
+	Mode   AccessMode `gorm:"NOT NULL"`
+}
+
+// AccessMode is the access mode of a user has to a repository.
+type AccessMode int
+
+const (
+	AccessModeNone  AccessMode = iota // 0
+	AccessModeRead                    // 1
+	AccessModeWrite                   // 2
+	AccessModeAdmin                   // 3
+	AccessModeOwner                   // 4
+)
+
+func (mode AccessMode) String() string {
+	switch mode {
+	case AccessModeRead:
+		return "read"
+	case AccessModeWrite:
+		return "write"
+	case AccessModeAdmin:
+		return "admin"
+	case AccessModeOwner:
+		return "owner"
+	default:
+		return "none"
+	}
+}
+
+// ParseAccessMode returns corresponding access mode to given permission string.
+func ParseAccessMode(permission string) AccessMode {
+	switch permission {
+	case "write":
+		return AccessModeWrite
+	case "admin":
+		return AccessModeAdmin
+	default:
+		return AccessModeRead
+	}
+}
+
 var _ PermsStore = (*perms)(nil)
 
 type perms struct {
 	*gorm.DB
 }
 
-func (db *perms) AccessMode(userID int64, repo *Repository) (mode AccessMode) {
-	if repo == nil {
+type AccessModeOptions struct {
+	OwnerID int64 // The ID of the repository owner.
+	Private bool  // Whether the repository is private.
+}
+
+func (db *perms) AccessMode(userID, repoID int64, opts AccessModeOptions) (mode AccessMode) {
+	if repoID <= 0 {
 		return AccessModeNone
 	}
 
 	// Everyone has read access to public repository.
-	if !repo.IsPrivate {
+	if !opts.Private {
 		mode = AccessModeRead
 	}
 
@@ -45,23 +98,23 @@ func (db *perms) AccessMode(userID int64, repo *Repository) (mode AccessMode) {
 		return mode
 	}
 
-	if userID == repo.OwnerID {
+	if userID == opts.OwnerID {
 		return AccessModeOwner
 	}
 
 	access := new(Access)
-	err := db.Where("user_id = ? AND repo_id = ?", userID, repo.ID).First(access).Error
+	err := db.Where("user_id = ? AND repo_id = ?", userID, repoID).First(access).Error
 	if err != nil {
 		if err != gorm.ErrRecordNotFound {
-			log.Error("Failed to get access [user_id: %d, repo_id: %d]: %v", userID, repo.ID, err)
+			log.Error("Failed to get access [user_id: %d, repo_id: %d]: %v", userID, repoID, err)
 		}
 		return mode
 	}
 	return access.Mode
 }
 
-func (db *perms) Authorize(userID int64, repo *Repository, desired AccessMode) bool {
-	return desired <= db.AccessMode(userID, repo)
+func (db *perms) Authorize(userID, repoID int64, desired AccessMode, opts AccessModeOptions) bool {
+	return desired <= db.AccessMode(userID, repoID, opts)
 }
 
 func (db *perms) SetRepoPerms(repoID int64, accessMap map[int64]AccessMode) error {
