+++ internal/db/ssh_key.go
@@ -753,10 +753,12 @@ func DeleteDeployKey(doer *User, id int64) error {
 		if err != nil {
 			return fmt.Errorf("GetRepositoryByID: %v", err)
 		}
-		yes, err := HasAccess(doer.ID, repo, AccessModeAdmin)
-		if err != nil {
-			return fmt.Errorf("HasAccess: %v", err)
-		} else if !yes {
+		if !Perms.Authorize(doer.ID, repo.ID, AccessModeAdmin,
+			AccessModeOptions{
+				OwnerID: repo.OwnerID,
+				Private: repo.IsPrivate,
+			},
+		) {
 			return ErrKeyAccessDenied{doer.ID, key.ID, "deploy"}
 		}
 	}
