+++ internal/db/testdata/backup/Access.golden.json
@@ -1,0 +1,2 @@
+{"ID":1,"UserID":1,"RepoID":11,"Mode":1}
+{"ID":2,"UserID":2,"RepoID":22,"Mode":2}
